Company 1 cannot be deleted any more.

+++ lib/Bloonix/Controller/Administration/Companies.pm
@@ -120,6 +120,10 @@ sub update {
 sub delete {
     my ($self, $c, $opts) = @_;
 
+    if ($opts->{id} == 1) {
+        return $c->plugin->error->no_privileges;
+    }
+
     $c->plugin->action->delete(company => $c->stash->object);
 }
